+++ .github/workflows/container.yml
@@ -153,9 +153,9 @@ jobs:
         shell: bash
         run: |
           ls -lR bpf
-          mkdir -p pkg/profiler/cpu/bpfprograms
-          cp -r bpf/out/* pkg/profiler/cpu/bpfprograms
-          rm pkg/profiler/cpu/bpfprograms/*/pid_namespace.bpf.o
+          mkdir -p pkg/profiler/cpu/bpf/programs
+          cp -r bpf/out/* pkg/profiler/cpu/bpf/programs
+          rm pkg/profiler/cpu/bpf/programs/*/pid_namespace.bpf.o
           mkdir -p pkg/contained/bpf
           cp -r bpf/out/* pkg/contained/bpf
           find pkg/contained/bpf/*/ -type f | grep -v "pid_namespace.bpf.o" | xargs -I{} bash -c "rm {}"
