+++ pkg/profiler/cpu/bpf/metrics/collector_test.go
