+++ pkg/profiler/cpu/bpf/programs/programs.go
